Overhaul HoundDawgs.tracker
Torrentname has now been fixed to include the entire torrentname (releasegroup was omitted, and captured in its own capturegroup to match uploader)
Added support for matching releasegroup (and uploader for backwards compatibility after a community vote, too many issues if not included I assume)
Split music from the main string to cope with weird torrentnames - thereby fixing issues with releasegroup matching
Added support for FreeLeech option

I tried to keep the RegEx as readable as possible. Some might argue that it could be shortened but from testing it revealed that shortening it would not decrease the amount of steps needed for a match and only led to the RegEx to be harder to understand.
Some might argue that I could use the (scene) tag in some of the music uploads but they far between and would only result in more errors.